+++ src/Aplication/LoanSimulation/DTOs/BankRateResult.cs
@@ -5,8 +5,8 @@
         //a api retorna os 8 primeiros digitos do cnpj
         public string? Cnpj { get; set; }
         public string? BankName { get; set; }
-        public decimal MonthlyRate { get; set; }
-        public decimal AnnualRate { get; set; }
+        public double MonthlyRate { get; set; }
+        public double AnnualRate { get; set; }
         public string? YearMonth { get; set; }
     }
 }
