+++ src/Infrastructure/Persistence/ApiResponse.cs
@@ -8,8 +8,8 @@
     public class LoanData
     {
         public string? InstituicaoFinanceira { get; set; }
-        public decimal TaxaJurosAoMes { get; set; }
-        public decimal TaxaJurosAoAno { get; set; }
+        public double TaxaJurosAoMes { get; set; }
+        public double TaxaJurosAoAno { get; set; }
         //a api retorna os 8 primeiros digitos do cnpj
         public string? Cnpj8 { get; set; }
         public string? AnoMes { get; set; }
